// src/app/models/product.ts
export interface Product {
  date: string;
  stockcode: string;
  product_name: string;
  price: number;
  is_on_special: boolean;
  is_half_price: boolean;
  was_price: number;
  savings_amount: number;
  package_size: string;
  unit_weight_in_grams: number;
  cup_price: number;
  cup_measure: string;
  cup_string: string;
  store: string;
}